bug fixed again for task 3, print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -30,7 +30,7 @@ void print_all(const char * const format, ...)
 				break;
 			case 'f':
 				arg = va_arg(ap, double);
-				printf("%d", arg);
+				printf("%f",(float)arg);
 				break;
 			case 's':
 				iarg = va_arg(ap, char *);
